requirements.txt
pydantic==1.8.2
python-telegram-bot==13.9
requests==2.26.0
tornado==6.1
